requirements.txt
# requirements.txt
# Pin versions used for training so the pickled pipeline loads correctly
fastapi==0.115.3
uvicorn[standard]==0.23.2
pandas==1.5.3
numpy==1.23.5
scikit-learn==1.1.3
joblib==1.2.0
sqlalchemy==2.0.29
psycopg2-binary==2.9.9
python-dotenv==1.0.1